feat: implement category management page with add, edit, and view modes
